+++ settings.example-bob.py
@@ -114,6 +114,10 @@ telemetry = {
 
 # Sensor configuration.
 sensors = {
+
+    # Whether to prettify sensor log output.
+    'prettify_log': True,
+
     'system': {
 
         # Adjust voltage divider resistor values matching the board.