+++ CHANGES.rst
@@ -54,6 +54,7 @@ Development
 - Desktop notifications for MicroTerkin Agent
 - Make MicroTerkin Agent write detected IP address into file
 - Improve Makefile sandbox tooling
+- Prettify logging
 
 
 2019-06-22 0.5.1